requirements.txt
appdirs==1.4.4
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.2
behave==1.2.6
behave-webdriver==0.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
gherkin-official==24.1.0
h11==0.14.0
idna==3.4
isodate==0.6.1
langdetect==1.0.9
outcome==1.2.0
parse==1.19.1
parse-type==0.6.2
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.13.0
selenium-utils==0.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.6
wsproto==1.2.0
